speaker_verification/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
speechbrain>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
captum>=0.6.0
tqdm>=4.65.0
scipy>=1.11.0
librosa>=0.10.0
soundfile>=0.12.0
PyYAML>=6.0
tensorboard>=2.13.0
pytest>=7.4.0
pathlib2>=2.3.7
